+++ resource_cache/resources/resource.hpp
@@ -11,9 +11,8 @@ class ResourceCache;
 class Resource
 {
 public:
-	Resource(std::string name, uint8_t* buffer, unsigned int size,
-		std::shared_ptr<IResourceStore> store,
-		std::shared_ptr<ResourceCache> cache);
+	Resource(std::string name, uint8_t* buffer, uint64_t size,
+		std::shared_ptr<IResourceStore> store, std::shared_ptr<ResourceCache> cache);
 
 	virtual ~Resource();
 
@@ -21,7 +20,7 @@ public:
 
 	[[nodiscard]] inline std::shared_ptr<IResourceStore> store() const { return _store; }
 	[[nodiscard]] inline std::string name() const { return _name; }
-	[[nodiscard]] inline unsigned int size() const { return _size; }
+	[[nodiscard]] inline uint64_t size() const { return _size; }
 	[[nodiscard]] inline uint8_t* buffer() const { return _buffer; }
 	[[nodiscard]] inline uint8_t* writableBuffer() const { return _buffer; }
 	[[nodiscard]] virtual ResourceType type() const { return ResourceType::General; }
@@ -29,7 +28,7 @@ public:
 protected:
 	std::string _name;
 	uint8_t* _buffer;
-	unsigned int _size;
+	uint64_t _size;
 	std::shared_ptr<IResourceStore> _store;
 	std::shared_ptr<ResourceCache> _cache;
 };
